Finish initial organization of Team class
