+++ lib/stat_tracker.rb
@@ -2,6 +2,7 @@ require 'csv'
 require './lib/csv_loader.rb'
 require './lib/games'
 require './lib/league'
+require './lib/team_stats'
 
 
 class StatTracker
@@ -10,6 +11,7 @@ class StatTracker
   def initialize (games, teams, game_teams)
     @gamezz = Games.new(games, teams, game_teams)
     @league = League.new(games, teams, game_teams)
+    @team_stat = Team.new(games, teams, game_teams)
     # @csv = CsvLoader.new(games, teams, game_teams)
     # @details = DetailsLoader.new.load_all
     # @games = games
@@ -367,226 +369,232 @@ class StatTracker
   end
 
   def team_info(team_id) #issue # 23 - Pass
+    @team_stat.team_info
+  #     info = {
+  #     "team_id" => team_id,
+  #     "franchise_id" => 0,
+  #     "team_name" => 0,
+  #     "abbreviation" => 0,
+  #     "link" => 0
+  #             }
 
-      info = {
-      "team_id" => team_id,
-      "franchise_id" => 0,
-      "team_name" => 0,
-      "abbreviation" => 0,
-      "link" => 0
-              }
+  #     @teams.each do |row|
+  #       info["franchise_id"] = row[:franchiseid].to_s if row[:team_id] == team_id.to_i
+  #       info["team_name"] = row[:teamname] if row[:team_id] == team_id.to_i
+  #       info["abbreviation"] = row[:abbreviation] if row[:team_id] == team_id.to_i
+  #       info["link"] = row[:link] if row[:team_id] == team_id.to_i
+  #     end
 
-      @teams.each do |row|
-        info["franchise_id"] = row[:franchiseid].to_s if row[:team_id] == team_id.to_i
-        info["team_name"] = row[:teamname] if row[:team_id] == team_id.to_i
-        info["abbreviation"] = row[:abbreviation] if row[:team_id] == team_id.to_i
-        info["link"] = row[:link] if row[:team_id] == team_id.to_i
-      end
-
-      info
+  #     info
   end
 
-  def games_by_season # All game_ids sorted by season - helper method for issue #18
-    # {20122013=> [2012030221, 2012030222, 2012030223, 2012030224,
-    games_by_season = {}
-    @games.values_at(:game_id, :season).each do |game|
-      if games_by_season.include?(game[1])
-        games_by_season[game[1]] << game[0]
-      else
-        games_by_season[game[1]] = [game[0]]
-      end
-    end
-    games_by_season
-  end
+  # def games_by_season # All game_ids sorted by season - helper method for issue #18
+    
+  #  # {20122013=> [2012030221, 2012030222, 2012030223, 2012030224,
+  #   games_by_season = {}
+  #   @games.values_at(:game_id, :season).each do |game|
+  #     if games_by_season.include?(game[1])
+  #       games_by_season[game[1]] << game[0]
+  #     else
+  #       games_by_season[game[1]] = [game[0]]
+  #     end
+  #   end
+  #   games_by_season
+  #  end
 
-  def wins_by_team(team_id) # List of every game that was a win for a team - helper method for issue #18
-    # [[2013020252, 16], [2014030166, 16], [2016030151, 16], [2016030152, 16]]
-    # wins = []
-    # @games.each do |row|
-    #   if (row[:away_goals] > row[:home_goals] && row[:away_team_id] == team_id) ||
-    #     (row[:home_goals] < row[:away_goals] && row[:home_team_id] == team_id)
-    #     wins << [row[:game_id], team_id]
-    #   end
-    # end
-    # wins
+  # def wins_by_team(team_id) # List of every game that was a win for a team - helper method for issue #18
+  #   # [[2013020252, 16], [2014030166, 16], [2016030151, 16], [2016030152, 16]]
+  #   # wins = []
+  #   # @games.each do |row|
+  #   #   if (row[:away_goals] > row[:home_goals] && row[:away_team_id] == team_id) ||
+  #   #     (row[:home_goals] < row[:away_goals] && row[:home_team_id] == team_id)
+  #   #     wins << [row[:game_id], team_id]
+  #   #   end
+  #   # end
+  #   # wins
 
-    # Option with full data set, but does not work with current dummy data
-    # (this could be made dynamic for win or loss):
-    #def results_by_team(team_id, win_loss)
-    result_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
-      game[1] == team_id && game[2] == "WIN"
-    end
-  end
+  #   # Option with full data set, but does not work with current dummy data
+  #   # (this could be made dynamic for win or loss):
+  #   #def results_by_team(team_id, win_loss)
+  #   result_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
+  #     game[1] == team_id && game[2] == "WIN"
+  #   end
+  # end
 
-  def games_by_team(team_id) # List of every game a team played - helper method for issue #18
-    # [[2013020252, 16], [2013020987, 16], [2014020903, 16], [2012020574, 16], [2014030161, 16],
-    games = []
-    @game_teams.each do |row|
-      games <<[row[:game_id], team_id] if (row[:team_id] == team_id)
-    end
-    games
+  # def games_by_team(team_id) # List of every game a team played - helper method for issue #18
+  #   # [[2013020252, 16], [2013020987, 16], [2014020903, 16], [2012020574, 16], [2014030161, 16],
+  #   games = []
+  #   @game_teams.each do |row|
+  #     games <<[row[:game_id], team_id] if (row[:team_id] == team_id)
+  #   end
+  #   games
 
     # Option with full data set, but does not work with current dummy data:
     # games_by_team = @game_teams.values_at(:game_id, :team_id, :result).find_all do |game|
     #   game[1] == team_id
     # end
+  # end
+
+  # def number_team_games_per_season(team_id) # Count of number of games a team played each season - helper method for issue #18
+  #   # {20122013=>1, 20132014=>2, 20142015=>7, 20162017=>4}
+  #   team_games_by_season = Hash.new{0}
+  #   games_by_season.each do |season, games|
+  #     games_by_team(team_id).each do |result_data|
+  #       if games.include?(result_data[0])
+  #         team_games_by_season[season] += 1.0
+  #       elsif team_games_by_season[season] == 0
+  #         team_games_by_season[season] = 0.0
+  #       end
+  #     end
+  #   end
+  #   team_games_by_season
+  # end
+
+  # def number_team_wins_per_season(team_id) # Count of number of games a team won each season -helper method for issue #18
+  #   # {20132014=>1, 20142015=>1, 20162017=>2}
+  #   wins_by_season = Hash.new{0}
+  #   games_by_season.each do |season, games|
+  #     wins_by_team(team_id).each do |result_data|
+  #       if games.include?(result_data[0])
+  #         wins_by_season[season] += 1.0
+  #       end
+  #     end
+  #     if wins_by_season[season] == 0
+  #       wins_by_season[season] = 0.0
+  #     end
+  #   end
+  #   wins_by_season
+  # end
+
+  # def season_win_percentage(team_id) # Percentage of won games per season by team - helper method for issue #18
+  #   # {20132014=>50.0, 20142015=>14.3, 20162017=>50.0}
+  #   win_percentage = Hash.new
+  #   number_team_games_per_season(team_id).each do |game_season, game_count|
+  #     number_team_wins_per_season(team_id).each do |wins_season, win_count|
+  #       if game_season == wins_season
+  #         if win_count == 0
+  #           percentage = 0.0
+  #         else
+  #           percentage = ((win_count/game_count.to_f) * 100).round(1)
+  #         end
+  #         win_percentage[game_season] = percentage
+  #       end
+  #     end
+  #   end
+  #   win_percentage
+  # end
+
+   def best_season(team_id) #issue # 18 - PASS
+   @team_stat.best_season(team_id)
+    # season_win_percentage(team_id.to_i).key(season_win_percentage(team_id.to_i).values.max).to_s
   end
 
-  def number_team_games_per_season(team_id) # Count of number of games a team played each season - helper method for issue #18
-    # {20122013=>1, 20132014=>2, 20142015=>7, 20162017=>4}
-    team_games_by_season = Hash.new{0}
-    games_by_season.each do |season, games|
-      games_by_team(team_id).each do |result_data|
-        if games.include?(result_data[0])
-          team_games_by_season[season] += 1.0
-        elsif team_games_by_season[season] == 0
-          team_games_by_season[season] = 0.0
-        end
-      end
-    end
-    team_games_by_season
-  end
-
-  def number_team_wins_per_season(team_id) # Count of number of games a team won each season -helper method for issue #18
-    # {20132014=>1, 20142015=>1, 20162017=>2}
-    wins_by_season = Hash.new{0}
-    games_by_season.each do |season, games|
-      wins_by_team(team_id).each do |result_data|
-        if games.include?(result_data[0])
-          wins_by_season[season] += 1.0
-        end
-      end
-      if wins_by_season[season] == 0
-        wins_by_season[season] = 0.0
-      end
-    end
-    wins_by_season
-  end
-
-  def season_win_percentage(team_id) # Percentage of won games per season by team - helper method for issue #18
-    # {20132014=>50.0, 20142015=>14.3, 20162017=>50.0}
-    win_percentage = Hash.new
-    number_team_games_per_season(team_id).each do |game_season, game_count|
-      number_team_wins_per_season(team_id).each do |wins_season, win_count|
-        if game_season == wins_season
-          if win_count == 0
-            percentage = 0.0
-          else
-            percentage = ((win_count/game_count.to_f) * 100).round(1)
-          end
-          win_percentage[game_season] = percentage
-        end
-      end
-    end
-    win_percentage
-  end
-
-  def best_season (team_id) #issue # 18 - PASS
-    season_win_percentage(team_id.to_i).key(season_win_percentage(team_id.to_i).values.max).to_s
-  end
-
-  def worst_season(team_id) #issue # 25 - Fail due to not written
-    season_win_percentage(team_id.to_i).key(season_win_percentage(team_id.to_i).values.min).to_s
+  def worst_season(team_id)#issue # 25 - Fail due to not written
+    @team_stat.worst_season(team_id)
+    # season_win_percentage(team_id.to_i).key(season_win_percentage(team_id.to_i).values.min).to_s
   end
 
   def average_win_percentage(team_id) #issue # 20
-    (wins_by_team(team_id.to_i).count.to_f/games_by_team(team_id.to_i).count.to_f).round(2)
+    @team_stat.average_win_percentage(team_id)
+    # (wins_by_team(team_id.to_i).count.to_f/games_by_team(team_id.to_i).count.to_f).round(2)
   end
 
   def most_goals_scored(team_id) #issue # 27 pass
+    @team_stat.most_goals_scored(team_id)
+    # array_of_goals_for_specified_team = []
 
-    array_of_goals_for_specified_team = []
+    # @games.each do |row|
+    #   array_of_goals_for_specified_team << row[:away_goals] if team_id.to_i == row[:away_team_id]
+    #   array_of_goals_for_specified_team << row[:home_goals] if team_id.to_i == row[:home_team_id]
+    # end
 
-    @games.each do |row|
-      array_of_goals_for_specified_team << row[:away_goals] if team_id.to_i == row[:away_team_id]
-      array_of_goals_for_specified_team << row[:home_goals] if team_id.to_i == row[:home_team_id]
-    end
-
-    array_of_goals_for_specified_team.max()
+    # array_of_goals_for_specified_team.max()
   end
 
   def fewest_goals_scored(team_id) #issue # 28 - Fail due to not written
+    @team_stat.fewest_goals_scored(team_id)
+    # array_of_goals_for_specified_team = []
 
-    array_of_goals_for_specified_team = []
+    #   @games.each do |row|
+    #     array_of_goals_for_specified_team << row[:away_goals] if team_id.to_i == row[:away_team_id]
+    #     array_of_goals_for_specified_team << row[:home_goals] if team_id.to_i == row[:home_team_id]
+    #   end
 
-      @games.each do |row|
-        array_of_goals_for_specified_team << row[:away_goals] if team_id.to_i == row[:away_team_id]
-        array_of_goals_for_specified_team << row[:home_goals] if team_id.to_i == row[:home_team_id]
-      end
-
-      array_of_goals_for_specified_team.min()
+    #   array_of_goals_for_specified_team.min()
   end
 
   #issue # 29 - Fail due to not written
   def favorite_opponent(team_id)#issue # 29
-    rival_opp = {}
-    rival_opp_wins = rival_wins(team_id)
-    rival_opp_games = rival_game(team_id)
-    rival_opp_games.each do | rogk, rogv |
-      rival_opp_wins.each do | rowk, rowv |
-        if rogk == rowk
-          rival_opp.merge!("#{rowk}" => (rowv.to_f / rogv.to_f))
-        elsif rival_opp_wins[rogk].nil?
-          rival_opp[rogk.to_s] = 0.0
-        end
-      end
-    end
-    rival_opp.each do |k, v|
-      if v == rival_opp.values.min
-        return team_by_id[k.to_i]
-      end
-    end
+    @team_stat.favorite_opponent(team_id)
+    #   rival_opp = {}
+  #   rival_opp_wins = rival_wins(team_id)
+  #   rival_opp_games = rival_game(team_id)
+  #   rival_opp_games.each do | rogk, rogv |
+  #     rival_opp_wins.each do | rowk, rowv |
+  #       if rogk == rowk
+  #         rival_opp.merge!("#{rowk}" => (rowv.to_f / rogv.to_f))
+  #       elsif rival_opp_wins[rogk].nil?
+  #         rival_opp[rogk.to_s] = 0.0
+  #       end
+  #     end
+  #   end
+  #   rival_opp.each do |k, v|
+  #     if v == rival_opp.values.min
+  #       return team_by_id[k.to_i]
+  #     end
+  #   end
   end
 
-  def rival_wins(team_id) #helper for #24 and possibly fave opp
-    rivals_wins = []
-    @games.each do |row|
-      if row[:away_team_id] == team_id.to_i
-        if row[:away_goals] < row[:home_goals]
-          rivals_wins << row[:home_team_id]
-        end
-      end
-      if row[:home_team_id] == team_id.to_i
-        if row[:home_goals] < row[:away_goals]
-          rivals_wins << row[:away_team_id]
-        end
-      end
-    end
-      rivals_wins_hash = rivals_wins.tally
-      rivals_wins_hash
-  end
+  # def rival_wins(team_id) #helper for #24 and possibly fave opp
+  #   rivals_wins = []
+  #   @games.each do |row|
+  #     if row[:away_team_id] == team_id.to_i
+  #       if row[:away_goals] < row[:home_goals]
+  #         rivals_wins << row[:home_team_id]
+  #       end
+  #     end
+  #     if row[:home_team_id] == team_id.to_i
+  #       if row[:home_goals] < row[:away_goals]
+  #         rivals_wins << row[:away_team_id]
+  #       end
+  #     end
+  #   end
+  #     rivals_wins_hash = rivals_wins.tally
+  #     rivals_wins_hash
+  # end
 
-  def rival_game(team_id) #helper for #24 and possibly fave opp
-    rivals_games = []
-    games.each do |row|
-      if row[:home_team_id] == team_id.to_i
-      rivals_games << row[:away_team_id]
-      end
-      if row[:away_team_id] == team_id.to_i
-      rivals_games << row[:home_team_id]
-      end
-    end
-    rivals_games_hash = rivals_games.tally
-    rivals_games_hash
-  end
+  # def rival_game(team_id) #helper for #24 and possibly fave opp
+  #   rivals_games = []
+  #   games.each do |row|
+  #     if row[:home_team_id] == team_id.to_i
+  #     rivals_games << row[:away_team_id]
+  #     end
+  #     if row[:away_team_id] == team_id.to_i
+  #     rivals_games << row[:home_team_id]
+  #     end
+  #   end
+  #   rivals_games_hash = rivals_games.tally
+  #   rivals_games_hash
+  # end
 
   def rival(team_id) #issue 24 - Fail due to not written
-    rival_opp = {}
-    rival_opp_wins = rival_wins(team_id)
-    rival_opp_games = rival_game(team_id)
-    rival_opp_games.each do | rogk, rogv |
-      rival_opp_wins.each do | rowk, rowv |
-        if rogk == rowk
-          rival_opp.merge!("#{rowk}" => (rowv.to_f / rogv.to_f))
-        elsif rival_opp_wins[rogk].nil?
-          rival_opp[rogk.to_s] = 0.0
-        end
-      end
-    end
-    rival_opp.each do |k, v|
-      if v == rival_opp.values.max
-        return team_by_id[k.to_i]
-      end
-    end
+    @team_stat.rival(team_id)
+    #   rival_opp = {}
+  #   rival_opp_wins = rival_wins(team_id)
+  #   rival_opp_games = rival_game(team_id)
+  #   rival_opp_games.each do | rogk, rogv |
+  #     rival_opp_wins.each do | rowk, rowv |
+  #       if rogk == rowk
+  #         rival_opp.merge!("#{rowk}" => (rowv.to_f / rogv.to_f))
+  #       elsif rival_opp_wins[rogk].nil?
+  #         rival_opp[rogk.to_s] = 0.0
+  #       end
+  #     end
+  #   end
+  #   rival_opp.each do |k, v|
+  #     if v == rival_opp.values.max
+  #       return team_by_id[k.to_i]
+  #     end
+  #   end
   end
 end
